+++ src/components/BackgroundPicker.tsx
@@ -17,6 +17,7 @@ import { IBackgroundBookmark } from "../types/IBackgroundBookmark";
 // import { fuzzy } from "fast-fuzzy";
 
 interface IBackgroundList {
+    update: string;
     background: {
         [key: string]: string[];
     };
@@ -89,6 +90,9 @@ const BackgroundPicker: React.FC<BackgroundPickerProps> = ({
             setFunction(`/background_compressed/${bg}.jpg`);
 
             setFilterValue("all");
+            if (bg.includes("bg_s000362")) {
+                setErrorInformation("Give her a break. She's happy now.");
+            }
         } catch (error) {
             setErrorInformation(String(error));
             console.error(error);
@@ -127,6 +131,7 @@ const BackgroundPicker: React.FC<BackgroundPickerProps> = ({
 
     const renderBackgroundType = (type: string) => {
         if (type === "bookmarks" && backgroundBookmarks.length === 0) return;
+        if (filterValue === "all" && type === "cards") return;
 
         return (
             <div
@@ -134,11 +139,7 @@ const BackgroundPicker: React.FC<BackgroundPickerProps> = ({
                 key={type}
             >
                 <div className="width-100 center text-center">
-                    <h1 className="white">
-                        {background?.filename.includes("Kisaragi")
-                            ? ""
-                            : t(`group.${type}`)}
-                    </h1>
+                    <h1 className="white">{t(`group.${type}`)}</h1>
                 </div>
                 <div className="flex-wrap center width-100">
                     {type === "bookmarks"
@@ -167,11 +168,7 @@ const BackgroundPicker: React.FC<BackgroundPickerProps> = ({
                                           ? "picker-item-selected"
                                           : ""
                                   }`}
-                                  src={
-                                      background?.filename.includes("Kisaragi")
-                                          ? `/img/transparent.png`
-                                          : `/background_low_jpg/${bg}.jpg`
-                                  }
+                                  src={`/background_low_jpg/${bg}.jpg`}
                                   onClick={async () => {
                                       handleChangeBackground(bg);
                                       setShow(false);
@@ -205,17 +202,15 @@ const BackgroundPicker: React.FC<BackgroundPickerProps> = ({
                         }}
                         value={filterValue}
                     >
-                        <option value="all">
-                            {background?.filename.includes("Kisaragi")
-                                ? ""
-                                : t("group.all")}
-                        </option>
+                        <option value="all">{t("group.all")}</option>
                         {Object.keys(backgroundList.background).map((type) => {
                             return (
                                 <option key={type} value={type}>
-                                    {background?.filename.includes("Kisaragi")
-                                        ? ""
-                                        : t(`group.${type}`)}
+                                    {`${t(`group.${type}`)} ${
+                                        type === "cards"
+                                            ? `(${backgroundList.update})`
+                                            : ""
+                                    }`}
                                 </option>
                             );
                         })}